Without isolation
Without a transformer, the working principle of theresonant converter is made clear. Changing Lr changes the Q value, which determines the sharpness of the decrease of the voltage gain with respect to switching frequency decrease. R is fixed to 6 ohms.
It is observed that it is very much in our hands to change Lr-fsw.  We donot need to change fsw too much, hence affect the impedance of the capacitor/ inductors.