Report #2
I tried to sweep the frequency keeping it in 50kHz range. I increased the q_e value accordingly with increasing fsuch that Lr is 100 uh always. What I have observed is that this does not affect the least voltage value, does not shrink the voltage range. So there is no way to decrease the output voltage to 0 without changing the frequency too much. (Bear in mind that Lr is kept constant)